api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
torch>=2.1.0
torchvision>=0.16.0
pytorch-grad-cam>=1.5.0
Pillow>=10.0.0
numpy>=1.24.0
